# api/openapi.json
{
  "openapi": "3.0",
  "info": {
    "title": "Sparvnest",
    "version": "0.1"
  },
  "components": {
    "securitySchemes": {
      "jwt_key": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
    "/login": {
      "post": {
        "description": "Logs in user into application",
        "requestBody": {
          "description": "Sending user email and password to server to authenticate user",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful authentication and login of user into application",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "relation_id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "forename": {
                      "type": "string"
                    },
                    "surname": {
                      "type": "string"
                    },
                    "email": {
                      "type": "string",
                      "format": "email"
                    },
                    "token": {
                      "type": "string",
                      "format": "jwt"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "The request cannot be computed - client sent invalid data to server endpoint",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "User authentication failed. Invalid authorization credentials",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Something went wrong on the server side - Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/users": {
      "post": {
        "description": "Registering an user (coach) to the system",
        "requestBody": {
          "description": "Sending needed data for user registration",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "password": {
                    "type": "string"
                  },
                  "forename": {
                    "type": "string"
                  },
                  "surname": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User successful registered",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "email": {
                      "type": "string",
                      "format": "email"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "The request cannot be computed - client sent invalid data to server endpoint",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "User was not registered because there is an user with the same email available",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Something went wrong on the server side - Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "description": "Getting profile information from currently logged in user (coach)",
        "security": {
          "jwt_key": []
        },
        "requestBody": {
          "description": "Sending authentication information to get profile information linked to userId",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "relation_id": {
                    "type": "string",
                    "format": "uuid"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User is authenticated and user profile information were gathered from database",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "forename": {
                      "type": "string"
                    },
                    "surname": {
                      "type": "string"
                    },
                    "email": {
                      "type": "string",
                      "format": "email"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "The request cannot be computed - client sent invalid data to server endpoint",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "User authentication failed. Invalid authorization credentials",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Something went wrong on the server side - Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "description": "Updating profile information of current logged in user",
        "security": {
          "jwt_key": []
        },
        "requestBody": {
          "description": "sending user profile information which should be updated - if anything shouldnt be updated, dont send it",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "relation_id": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "forename": {
                    "type": "string"
                  },
                  "surname": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User authentication was verified and user profile information are updated based on the data sent in the request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "The request cannot be computed - client sent invalid data to server endpoint",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "User authentication failed. Invalid authorization credentials",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Something went wrong on the server side - Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }

        }
      },
      "delete": {
        "description": "deleting currently logged in user",
        "security": {
          "jwt_key": []
        },
        "requestBody": {
          "description": "sending user profile information which should be updated - if anything shouldnt be updated, dont send it",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "relation_id": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User authentication was verified via token and deletion process was confirmed by sending the password as string of current user.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "The request cannot be computed - client sent invalid data to server endpoint",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "User authentication failed. Invalid authorization credentials or invalid email confirmation string",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Something went wrong on the server side - Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}